requirements.txt
viam-sdk
ikpy
smbus2
